+++ src/AppNxtp.tsx
@@ -4,7 +4,7 @@ import './App.css';
 import Web3ConnectionManager from './components/web3/Web3ConnectionManager';
 import WrappedWeb3ReactProvider from './components/web3/WrappedWeb3ReactProvider';
 import analytics from './services/analytics';
-import SwapNxtp from './components/SwapNxtp';
+import SwapNxtp from './components/nxtpDemo/SwapNxtp';
 
 function usePageViews() {
   let location = (window as any).location
